backend/oauth2/oauth2.py
from datetime import datetime, timedelta, timezone
from typing import Annotated, List, Optional

import jwt
from fastapi import Depends, HTTPException, status
from fastapi.security import OAuth2PasswordBearer
from passlib.context import CryptContext
from sqlalchemy.orm import Session
from api.utils.database import get_db
from database.users.role import Role
from oauth2.oauth2_model import TokenData
from database.users.user import User 
from services import user_service


SECRET_KEY = "ca65668bc0eb49efc4df1c8437c46a7b73b85e13419fa33564939651cbd5b2c4"
ALGORITHM = "HS256"
ACCESS_TOKEN_EXPIRE_MINUTES = 30

pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")
oauth2_scheme = OAuth2PasswordBearer(tokenUrl="token")




def verify_password(plain_password, hashed_password):
    return pwd_context.verify(plain_password, hashed_password)


def authenticate_user(db: Session, user_id: str, password: str) -> User | None:
    # Fetch the user by user_id
    user = user_service.get_user_byId(db, user_id)
    if not user or not user.profile:  # Check if the user and profile exist
        return False
    # Verify the password using the `profile`'s password attribute
    if not verify_password(password, user.profile.password):
        return False
    return user



# def authenticate_user(db: Session, user_id: str, password: str)-> User | None:
#     user = user_service.get_user_byId(db, user_id)
#     if not user or not verify_password(password, user.password):
#         return False
#     return user

def create_access_token(data: dict, expires_delta: Optional[timedelta] = None):
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + (expires_delta if expires_delta else timedelta(minutes=15))
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
    return encoded_jwt

async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)], db: Session = Depends(get_db)):
    credentials_exception = HTTPException(
        status_code=status.HTTP_401_UNAUTHORIZED,
        detail="Could not validate credentials",
        headers={"WWW-Authenticate": "Bearer"},
    )
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        user_id: str = payload.get("sub")
        if user_id is None:
            raise credentials_exception
        
        token_data = TokenData(user_id=user_id)
    except jwt.PyJWTError:
        raise credentials_exception

    user = user_service.get_user_byId(db, user_id=token_data.user_id)
    if not user:
        raise credentials_exception
    return user

async def get_current_active_user(current_user: Annotated[User, Depends(get_current_user)]):
    # Fetch profile from the database to check if the user is active
    profile = current_user.profile
    
    if profile is None or not profile.is_active:
        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Inactive user")
    return current_user




class RoleChecker:
    def __init__(self, allowed_roles: List):
        self.allowed_roles = allowed_roles

    def __call__(self, user: User = Depends(get_current_active_user)):
        if user.role not in self.allowed_roles:
            raise HTTPException(status_code=403, detail="Operation not permitted")


allow_admin = RoleChecker([Role])
